% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dataset_part2}
\alias{dataset_part2}
\title{Dataset for Section 2 of the Quantitative Analysis}
\format{
A data frame with 1691 rows and 8 variables:
\describe{
\item{ltla21_code}{\code{character} Local Authority code.}
\item{ltla21_name}{\code{character} Local Authority name.}
\item{year}{\code{double} Financial year ending, e.g. 2016 corresponds to a April 2015-March 2016 data collection period.}
\item{UC_households_perc}{\code{double} Percentage of households receiving Universal Credit.}
\item{children_low_income_perc}{\code{double} Percentage of children living in low-income families.}
\item{unemployment_perc}{\code{double} Unemployment rate in the local authority, expressed as a percentage.}
\item{months_active}{\code{double} Number of months Universal Credit has been active in the local authority.}
\item{years_active}{\code{double} Number of years Universal Credit has been active in the local authority.}
}
}
\usage{
dataset_part2
}
\description{
This dataset, second part of a series used for modeling analyses, extends \code{dataset_part1} by adding information about the duration of Universal Credit
activation within each local authority.
}
\examples{
data(dataset_part2)
summary(dataset_part2)
}
\keyword{datasets}
