% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{households_number_ltla}
\alias{households_number_ltla}
\title{Number of Households by Local Authority (2015-2020)}
\format{
A data frame with 2220 rows and 3 variables:
\describe{
\item{ltla21_code}{\code{character} Local Authority code.}
\item{year}{\code{double} Financial year ending, e.g. 2015 corresponds to a April 2014-March 2015 data collection period.}
\item{households_number}{\code{double} Number of households.}
}
}
\source{
https://www.ons.gov.uk/peoplepopulationandcommunity/birthsdeathsandmarriages/families/adhocs/13586estimatednumberofhouseholdsbyselectedhouseholdtypeslocalauthoritiesinenglandandwalescountiesandregionsofenglandscottishcouncilareasandgreatbritainconstituentcountries2004to2019

https://www.ons.gov.uk/peoplepopulationandcommunity/birthsdeathsandmarriages/families/adhocs/14432estimatednumberofhouseholdsingreatbritaininnuts1nuts3englishandwelshlocalauthoritiesandscottishcouncilareas2020

description
}
\usage{
households_number_ltla
}
\description{
This dataset contains the annual number of households within each local authority in England, Wales and Scotland.
}
\examples{
data(households_number_ltla)
summary(households_number_ltla)
}
\keyword{datasets}
