% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lone_parent_households_ltla}
\alias{lone_parent_households_ltla}
\title{Lone Parent Households by Local Authority (2015-2019)}
\format{
A data frame with 1826 rows and 6 variables:
\describe{
\item{ltla21_name}{\code{character} Local Authority (LTLA) name.}
\item{ltla21_code}{\code{character} Local Authority code.}
\item{year}{\code{double} Financial year ending, e.g. 2015 corresponds to a April 2014-March 2015 data collection period.}
\item{lone_parent_households_abs}{\code{double} Absolute number of lone parent households.}
\item{households_number}{\code{double} Total number of households in the LTLA.}
\item{lone_parent_households_perc}{\code{double} Percentage of households that are lone parent households.}
}
}
\source{
https://www.ons.gov.uk/peoplepopulationandcommunity/birthsdeathsandmarriages/families/adhocs/13586estimatednumberofhouseholdsbyselectedhouseholdtypeslocalauthoritiesinenglandandwalescountiesandregionsofenglandscottishcouncilareasandgreatbritainconstituentcountries2004to2019
}
\usage{
lone_parent_households_ltla
}
\description{
This dataset provides annual data on the number and proportion of lone parent households within each
local authority in England, Wales and Scotkand.
}
\examples{
data(lone_parent_households_ltla)
summary(lone_parent_households_ltla)
}
\keyword{datasets}
