% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_variable_ltla.R
\name{map_variable_ltla}
\alias{map_variable_ltla}
\title{Map a Variable at the Local Authority Level}
\usage{
map_variable_ltla(df, fill_variable)
}
\arguments{
\item{df}{DataFrame containing the data to be plotted, which must include a region identifier that matches that of the 'boundaries_ltla21' dataset of the geographr package (ltla21_code or ltla21_name).}

\item{fill_variable}{The name of the variable to be used for the fill aesthetic in the map. This must be provided as a string. For example, 'population_density'.}
}
\value{
A ggplot object representing the map with the specified variable highlighted.
}
\description{
This function takes a dataset and a specified variable from that dataset to create a
geographic map displaying the values of that variable across UK Local Authorities.
The boundaries are defined at the Lower Tier Local Authority (LTLA) level.
}
