% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pre_treatment_df}
\alias{pre_treatment_df}
\title{Pre-Treatment Data Frame for Local Authorities}
\format{
A data frame with 359 rows and 9 variables:
\describe{
\item{ltla21_name}{\code{character} Local Authority (LTLA) name.}
\item{ltla21_code}{\code{character} Local Authority code.}
\item{year}{\code{double} Pre-treatment year: 2015 for all records.}
\item{children_low_income_perc}{\code{double} Percentage of children living in low-income families.}
\item{households_number}{\code{double} Total number of households in the local authority.}
\item{lone_parent_households_perc}{\code{double} Percentage of households that are lone parent households.}
\item{unemployment_perc}{\code{double} Unemployment rate in the local authority, expressed as a percentage.}
\item{first_active_date}{\code{date} The date when Universal Credit first became active in the local authority.}
\item{first_active_quarter}{\code{factor} The quarter when Universal Credit first became active in the local authority.}
}
}
\usage{
pre_treatment_df
}
\description{
This dataset combines values for all variables used in the study for each local authority in England, Scotland and Wales
for the year 2015. It serves as a baseline for the evaluation of the impact of UC, which was introduced after this year.
}
\examples{
data(pre_treatment_df)
summary(pre_treatment_df)
}
\keyword{datasets}
