% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{uc_first_active_month}
\alias{uc_first_active_month}
\title{First Active Month of Universal Credit by Local Authority}
\format{
A data frame with 358 rows and 2 variables:
\describe{
\item{ltla21_code}{\code{character} Local Authority code.}
\item{first_active_date}{\code{date} Month when Universal Credit first became active in the local authority.}
}
}
\source{
https://assets.publishing.service.gov.uk/media/5ab507c8e5274a1aa2d414d1/universal-credit-transition-rollout-schedule.pdf
}
\usage{
uc_first_active_month
}
\description{
This dataset records the month when Universal Credit was introduced
in each local authority in England, Scotland and Wales. This information is useful for tracking the
rollout of Universal Credit across the UK.
}
\examples{
data(uc_first_active_month)
summary(uc_first_active_month)
}
\keyword{datasets}
