% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{UC_households_ltla}
\alias{UC_households_ltla}
\title{Households Claiming Universal Credit by Local Authority (2016-2020)}
\format{
A data frame with 1750 rows and 6 variables:
\describe{
\item{ltla21_name}{\code{character} Local Authority (LTLA) name.}
\item{ltla21_code}{\code{character} Local Authority code.}
\item{year}{\code{double} Financial year ending, e.g. 2016 corresponds to a April 2015-March 2016 data collection period.}
\item{UC_households_abs}{\code{double} Absolute number of households claiming Universal Credit.}
\item{households_number}{\code{double} Total number of households in the LTLA.}
\item{UC_households_perc}{\code{double} Percentage of households claiming Universal Credit in the LTLA.}
}
}
\source{
“Households on Universal Credit” dataset from Stat-Xplore
}
\usage{
UC_households_ltla
}
\description{
This dataset contains annual data on the number and proportion of households receiving Universal Credit (UC)
within each local authority in England, Scotland and Wales.
}
\examples{
data(UC_households_ltla)
summary(UC_households_ltla)
}
\keyword{datasets}
