% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{unemployment_ltla}
\alias{unemployment_ltla}
\title{Unemployment Rates by Local Authority (2015-2020)}
\format{
A data frame with 2220 rows and 4 variables:
\describe{
\item{ltla21_name}{\code{character} Local Authority (LTLA) name.}
\item{ltla21_code}{\code{character} Local Authority code.}
\item{year}{\code{double} Financial year ending, e.g. 2015 corresponds to a April 2014-March 2015 data collection period.}
\item{unemployment_perc}{\code{double} Unemployment rate in the local authority, expressed as a percentage.}
}
}
\source{
https://www.ons.gov.uk/employmentandlabourmarket/peoplenotinwork/unemployment/datasets/modelledunemploymentforlocalandunitaryauthoritiesm01/current
}
\usage{
unemployment_ltla
}
\description{
This dataset provides annual unemployment rates for each local authority in England, Scotland and Wales.
}
\examples{
data(unemployment_ltla)
summary(unemployment_ltla)
}
\keyword{datasets}
